requirements.txt
# Manually listed requirements
# albumentations==1.1.0
# h5py==3.8.0
# huggingface-hub==0.13.0
# kaleido==0.2.1
# mmcv==1.7.1
# mmsegmentation==0.30.0
# opencv-python-headless==4.7.0.68
# opencv-python==4.7.0.68
# pypdf2==1.26.0
# pytorch-lightning==1.8.6
# scikit-learn==1.0.2
# segmentation-models-pytorch==0.3.2
# timm==0.8.6
# torch==1.10.0
# torchmetrics==0.10.2
# torchvision==0.11.1
# black
# Cython
# CairoSVG
# debugpy
# jupyter
# nibabel
# pandas
# plotly
# pydicom
# pylibjpeg[all]
# python-gdcm
# seaborn
# tabulate
# termcolor
# wandb==0.15.8
# lightly==1.4.15

## Pip freeze
addict==2.4.0
aenum==3.1.15
aiohttp==3.8.5
aiosignal==1.3.1
albumentations==1.1.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astunparse==1.6.3
async-timeout==4.0.2
asynctest==0.13.0
attrs==23.1.0
black==23.3.0
bleach==6.0.0
bokeh==2.4.3
brotlipy==0.7.0
cairocffi==1.6.1
CairoSVG==2.7.1
certifi==2021.10.8
charset-normalizer==3.2.0
click==8.1.6
cloudpickle==2.2.1
colorcet==3.0.1
comm==0.1.4
conda==4.10.3
conda-build==3.21.5
cssselect2==0.7.0
cycler==0.11.0
Cython==3.0.0
daal==2023.1.1
daal4py==2023.1.1
dask==2022.2.0
datashader==0.14.4
datashape==0.5.2
debugpy==1.6.7.post1
defusedxml==0.7.1
dm-tree==0.1.8
dnspython==2.1.0
docker-pycreds==0.4.0
efficientnet-pytorch==0.7.1
einops==0.6.1
entrypoints==0.4
exceptiongroup==1.1.2
fastjsonschema==2.18.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
gast==0.5.4
gitdb==4.0.10
GitPython==3.1.32
h5py==3.8.0
holoviews==1.16.2
huggingface-hub==0.13.0
hydra-core==1.3.2
imageio==2.31.1
importlib-metadata==6.7.0
importlib-resources==5.12.0
ipykernel==6.16.2
ipython-genutils==0.2.0
ipywidgets==8.1.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.17.3
jupyter==1.0.0
jupyter-client==7.4.9
jupyter-console==6.6.3
jupyter-core==4.12.0
jupyter-server==1.24.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
kaleido==0.2.1
kiwisolver==1.4.4
lightly==1.4.21
lightly-utils==0.0.2
lightning-bolts==0.6.0.post1
lightning-utilities==0.9.0
llvmlite==0.39.1
locket==1.0.0
Markdown==3.4.4
matplotlib==3.5.3
mistune==3.0.1
mkl-fft==1.3.1
mkl-service==2.4.0
mmcls==0.25.0
mmcv==1.7.1
multidict==6.0.4
multipledispatch==1.0.0
munch==4.0.0
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.6.0
nbformat==5.8.0
nest-asyncio==1.5.7
networkx==2.6.3
nibabel==4.0.2
notebook==6.5.5
notebook-shim==0.2.3
numba==0.56.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-dali-cuda110==1.28.0
olefile==0.46
omegaconf==2.3.0
onnx==1.14.1
opencv-python==4.7.0.68
opencv-python-headless==4.7.0.68
packaging==23.1
pandas==1.3.5
pandocfilters==1.5.0
panel==0.14.4
param==1.13.0
partd==1.4.1
pathspec==0.11.2
pathtools==0.1.2
Pillow==8.4.0
pkginfo==1.7.1
pkgutil-resolve-name==1.3.10
platformdirs==3.10.0
plotly==5.15.0
pretrainedmodels==0.7.4
prettytable==3.7.0
prometheus-client==0.17.1
promise==2.3
prompt-toolkit==3.0.39
protobuf==4.24.0
pycosat==0.6.3
pyct==0.5.0
pydantic==1.10.12
pydicom==2.4.2
pylibjpeg==1.4.0
pylibjpeg-libjpeg==1.3.4
pylibjpeg-openjpeg==1.3.2
pylibjpeg-rle==1.3.0
pynndescent==0.5.10
pyparsing==3.1.1
PyPDF2==1.26.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-etcd==0.4.5
python-gdcm==3.0.22
python-hostlist==1.22
pytorch-lightning==1.8.6
pytorch-toolbelt==0.5.0
pytz==2021.3
pyviz-comms==2.3.2
PyWavelets==1.3.0
PyYAML==5.4.1
pyzmq==24.0.1
qtconsole==5.4.3
QtPy==2.3.1
qudida==0.0.4
safetensors==0.4.0
scikit-image==0.19.3
scikit-learn==1.0.2
scikit-learn-intelex==2023.1.1
scipy==1.7.3
seaborn==0.12.2
segmentation-models-pytorch==0.3.3
Send2Trash==1.8.2
sentry-sdk==1.29.2
setproctitle==1.3.2
shortuuid==1.0.11
simple-slurm==0.2.6
smmap==5.0.0
sniffio==1.3.0
tabulate==0.9.0
tbb==2021.9.0
tenacity==8.2.2
tensorboardX==2.6.2
tensorflow-addons==0.19.0
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.9.2
tinycss2==1.2.1
tomli==2.0.1
toolz==0.12.0
torch==1.10.0
torchelastic==0.2.0
torchmetrics==0.10.2
torchtext==0.11.0
torchvision==0.11.1
tornado==6.2
traitlets==5.9.0
typed-ast==1.5.5
typeguard==4.1.2
typing-extensions==4.7.1
umap-learn==0.5.3
urllib3==1.26.16
wandb==0.15.8
webencodings==0.5.1
websocket-client==1.6.1
widgetsnbextension==4.0.8
xarray==0.20.2
yapf==0.40.1
yarl==1.9.2
zipp==3.15.0